Button Controlled Motors
Simple functional demo (as shown in video) of switch controlling motor rpm. Note that the dodgy connection to the motors (and subsequent grounding on the motor housing) is causing the motor's intermittent output. (If the button stops the motors completely they can struggle to start again). Hence why in the video I don't completely stop the motors.